【修改】1、分页接口查询错误的问题

+++ icezerocat-mybatismp/src/main/java/github/com/icezerocat/mybatismp/service/impl/BaseCurdServiceImpl.java
@@ -201,7 +201,7 @@ public class BaseCurdServiceImpl implements BaseCurdService {
             return build.setCode(HttpStatus.OK.value()).setCount(selectPage.getTotal()).setData(selectPage.getRecords()).complete();
         } else {
             List selectList = baseMapperObjectNoahService.getBaseMapper().selectList(wrapper);
-            return build.setCode(HttpStatus.OK.value()).setCount((long) selectList.size()).setData(selectList).complete();
+            return build.setCode(HttpStatus.OK.value()).setCount(selectList.size()).setData(selectList).complete();
         }
     }
